Vendor everything
The HEExTokenizer is crashing when TailWindFormatter is used
with a more recent version of LiveView because HEExTokenizer
got out of sync with the other code in LiveView it depends on.

The best solution is to ask the LiveView team to make
HTMLFormatter.tokenize/1 a public function.

In the meantime, vendoring everything else that HEExTokenizer needs
fixes the issue.